requirements.txt
asyncio
datetime
pyfiglet
playsound
discord.py-self
python3-anticaptcha